opol/stack/services/classification-service/requirements.txt
prefect
openai
redis
fastapi
sqlmodel
pgvector
pydantic
instructor
google-generativeai
jsonref
logfire
logfire[asyncpg]
opol==0.1.19
pytest
